# backend/docs/swagger.yaml
basePath: /api
definitions:
  model.AddUserRequest:
    properties:
      choregroup_name:
        type: string
      password:
        type: string
      user_role:
        description: Should be 'user' or 'admin'
        type: string
      username:
        type: string
    type: object
  model.CreateTaskRequest:
    properties:
      assigned_to_user_id:
        type: string
      points_reward:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
  model.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  model.LoginResponse:
    properties:
      choregroup_id:
        type: string
      role:
        type: string
      user_id:
        type: string
    type: object
  model.SignUpRequest:
    properties:
      choregroup_name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  model.StatisticsResponse:
    properties:
      cooperative_points:
        type: integer
      users:
        items:
          $ref: '#/definitions/model.User'
        type: array
    type: object
  model.Task:
    properties:
      assigned_to_user_id:
        type: string
      choregroup_id:
        type: string
      id:
        type: string
      points_reward:
        type: integer
      status:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  model.TaskSubmission:
    properties:
      created_at:
        type: string
      id:
        type: string
      status:
        type: string
      submitted_by:
        type: string
      task_id:
        type: string
    type: object
  model.UpdateSubmissionRequest:
    properties:
      action:
        description: '"approve" or "reject"'
        type: string
    type: object
  model.UpdateTaskRequest:
    properties:
      assigned_to_user_id:
        type: string
      points_reward:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
  model.User:
    properties:
      choregroup_id:
        type: string
      id:
        type: string
      points:
        type: integer
      role:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the backend server for the ChoreCraft application.
  title: ChoreCraft API
  version: "1.0"
paths:
  /choregroups/{choregroupID}/members:
    get:
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: ChoreGroup ID
        in: path
        name: choregroupID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get all members of a choregroup
      tags:
      - Users
  /choregroups/{choregroupID}/members/{userID}:
    delete:
      parameters:
      - description: Admin User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: ChoreGroup ID
        in: path
        name: choregroupID
        required: true
        type: string
      - description: User ID to delete
        in: path
        name: userID
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete a user
      tags:
      - Admin
  /choregroups/{choregroupID}/statistics:
    get:
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: ChoreGroup ID
        in: path
        name: choregroupID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.StatisticsResponse'
      security:
      - ApiKeyAuth: []
      summary: Get choregroup statistics (user points and cooperative points)
      tags:
      - Statistics
  /choregroups/{choregroupID}/submissions:
    get:
      parameters:
      - description: Admin User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: ChoreGroup ID
        in: path
        name: choregroupID
        required: true
        type: string
      - description: Filter by status (e.g., pending_approval, approved)
        in: query
        name: status
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.TaskSubmission'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List submissions for a choregroup
      tags:
      - Admin
  /choregroups/{choregroupID}/tasks:
    get:
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: ChoreGroup ID
        in: path
        name: choregroupID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Task'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List tasks for a choregroup
      tags:
      - Tasks
    post:
      consumes:
      - application/json
      description: Creates a new chore.
      parameters:
      - description: Admin User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: ChoreGroup ID
        in: path
        name: choregroupID
        required: true
        type: string
      - description: Task details
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/model.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Task'
      security:
      - ApiKeyAuth: []
      summary: Create a new task for a choregroup
      tags:
      - Tasks
  /choregroups/{choregroupID}/tasks/{taskID}:
    delete:
      parameters:
      - description: Admin User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: ChoreGroup ID
        in: path
        name: choregroupID
        required: true
        type: string
      - description: Task ID
        in: path
        name: taskID
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete a task
      tags:
      - Admin
    put:
      consumes:
      - application/json
      parameters:
      - description: Admin User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: ChoreGroup ID
        in: path
        name: choregroupID
        required: true
        type: string
      - description: Task ID
        in: path
        name: taskID
        required: true
        type: string
      - description: Task details
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/model.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update an existing task
      tags:
      - Admin
  /choregroups/{choregroupID}/tasks/{taskID}/status:
    put:
      parameters:
      - description: Admin User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: ChoreGroup ID
        in: path
        name: choregroupID
        required: true
        type: string
      - description: Task ID
        in: path
        name: taskID
        required: true
        type: string
      - description: Action to perform (approve or reject)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateSubmissionRequest'
      responses:
        "204":
          description: No Content
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Approve or reject a task by its ID
      tags:
      - Admin
  /choregroups/{choregroupID}/tasks/{taskID}/submit:
    post:
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: ChoreGroup ID
        in: path
        name: choregroupID
        required: true
        type: string
      - description: Task ID
        in: path
        name: taskID
        required: true
        type: string
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.TaskSubmission'
      security:
      - ApiKeyAuth: []
      summary: Submit a task for approval
      tags:
      - Tasks
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: User Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LoginResponse'
      summary: Log in a user
      tags:
      - Auth
  /signup:
    post:
      consumes:
      - application/json
      parameters:
      - description: Signup Details
        in: body
        name: signup
        required: true
        schema:
          $ref: '#/definitions/model.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.User'
      summary: Sign up a new choregroup and admin user
      tags:
      - Auth
  /users:
    post:
      consumes:
      - application/json
      parameters:
      - description: User and ChoreGroup Details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.AddUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.User'
      summary: Add a user to an existing choregroup
      tags:
      - Auth
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-User-ID
    type: apiKey
swagger: "2.0"